+++ openstack/blockstorage/v3/volumes/results.go
@@ -5,11 +5,8 @@ import (
 	"time"
 
 	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
-// Attachment represents a Volume Attachment record
 type Attachment struct {
 	AttachedAt   time.Time `json:"-"`
 	AttachmentID string    `json:"attachment_id"`
@@ -20,7 +17,6 @@ type Attachment struct {
 	VolumeID     string    `json:"volume_id"`
 }
 
-// UnmarshalJSON is our unmarshalling helper
 func (r *Attachment) UnmarshalJSON(b []byte) error {
 	type tmp Attachment
 	var s struct {
@@ -38,7 +34,6 @@ func (r *Attachment) UnmarshalJSON(b []byte) error {
 	return err
 }
 
-// Volume contains all the information associated with an OpenStack Volume.
 type Volume struct {
 	// Unique identifier for the volume.
 	ID string `json:"id"`
@@ -80,7 +75,6 @@ type Volume struct {
 	Multiattach bool `json:"multiattach"`
 }
 
-// UnmarshalJSON another unmarshalling function
 func (r *Volume) UnmarshalJSON(b []byte) error {
 	type tmp Volume
 	var s struct {
@@ -98,74 +92,4 @@ func (r *Volume) UnmarshalJSON(b []byte) error {
 	r.UpdatedAt = time.Time(s.UpdatedAt)
 
 	return err
-}
-
-// VolumePage is a pagination.pager that is returned from a call to the List function.
-type VolumePage struct {
-	pagination.LinkedPageBase
-}
-
-// IsEmpty returns true if a ListResult contains no Volumes.
-func (r VolumePage) IsEmpty() (bool, error) {
-	volumes, err := ExtractVolumes(r)
-	return len(volumes) == 0, err
-}
-
-func (r VolumePage) NextPageURL() (string, error) {
-	var s struct {
-		Links []golangsdk.Link `json:"volumes_links"`
-	}
-	err := r.ExtractInto(&s)
-	if err != nil {
-		return "", err
-	}
-	return golangsdk.ExtractNextURL(s.Links)
-}
-
-// ExtractVolumes extracts and returns Volumes. It is used while iterating over a volumes.List call.
-func ExtractVolumes(r pagination.Page) ([]Volume, error) {
-	var res []Volume
-	err := ExtractVolumesInto(r, &res)
-	return res, err
-}
-
-type commonResult struct {
-	golangsdk.Result
-}
-
-// Extract will get the Volume object out of the commonResult object.
-func (r commonResult) Extract() (*Volume, error) {
-	var res Volume
-	err = extract.Into(raw.Body, &res)
-	return &res, err
-}
-
-// ExtractInto converts our response data into a volume struct
-func (r commonResult) ExtractInto(v interface{}) error {
-	return r.Result.ExtractIntoStructPtr(v, "volume")
-}
-
-// ExtractVolumesInto similar to ExtractInto but operates on a `list` of volumes
-func ExtractVolumesInto(r pagination.Page, v interface{}) error {
-	return r.(VolumePage).Result.ExtractIntoSlicePtr(v, "volumes")
-}
-
-// CreateResult contains the response body and error from a Create request.
-type CreateResult struct {
-	commonResult
-}
-
-// GetResult contains the response body and error from a Get request.
-type GetResult struct {
-	commonResult
-}
-
-// UpdateResult contains the response body and error from an Update request.
-type UpdateResult struct {
-	commonResult
-}
-
-// DeleteResult contains the response body and error from a Delete request.
-type DeleteResult struct {
-	golangsdk.ErrResult
 }
