+++ openstack/blockstorage/v3/volumetypes/results.go
@@ -1,12 +1,5 @@
 package volumetypes
 
-import (
-	"github.com/opentelekomcloud/gophertelekomcloud"
-	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
-	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
-)
-
-// Volume Type contains all the information associated with an OpenStack Volume Type.
 type VolumeType struct {
 	// Unique identifier for the volume type.
 	ID string `json:"id"`
@@ -22,74 +15,4 @@ type VolumeType struct {
 	QosSpecID string `json:"qos_specs_id"`
 	// Volume Type access public attribute
 	PublicAccess bool `json:"os-volume-type-access:is_public"`
-}
-
-// VolumeTypePage is a pagination.pager that is returned from a call to the List function.
-type VolumeTypePage struct {
-	pagination.LinkedPageBase
-}
-
-// IsEmpty returns true if a ListResult contains no Volume Types.
-func (r VolumeTypePage) IsEmpty() (bool, error) {
-	volumetypes, err := ExtractVolumeTypes(r)
-	return len(volumetypes) == 0, err
-}
-
-func (r VolumeTypePage) NextPageURL() (string, error) {
-	var s struct {
-		Links []golangsdk.Link `json:"volume_type_links"`
-	}
-	err := r.ExtractInto(&s)
-	if err != nil {
-		return "", err
-	}
-	return golangsdk.ExtractNextURL(s.Links)
-}
-
-// ExtractVolumeTypes extracts and returns Volumes. It is used while iterating over a volumetypes.List call.
-func ExtractVolumeTypes(r pagination.Page) ([]VolumeType, error) {
-	var res []VolumeType
-	err := ExtractVolumeTypesInto(r, &res)
-	return res, err
-}
-
-type commonResult struct {
-	golangsdk.Result
-}
-
-// Extract will get the Volume Type object out of the commonResult object.
-func (r commonResult) Extract() (*VolumeType, error) {
-	var res VolumeType
-	err = extract.Into(raw.Body, &res)
-	return &res, err
-}
-
-// ExtractInto converts our response data into a volume type struct
-func (r commonResult) ExtractInto(v interface{}) error {
-	return r.Result.ExtractIntoStructPtr(v, "volume_type")
-}
-
-// ExtractVolumesInto similar to ExtractInto but operates on a `list` of volume types
-func ExtractVolumeTypesInto(r pagination.Page, v interface{}) error {
-	return r.(VolumeTypePage).Result.ExtractIntoSlicePtr(v, "volume_types")
-}
-
-// GetResult contains the response body and error from a Get request.
-type GetResult struct {
-	commonResult
-}
-
-// CreateResult contains the response body and error from a Create request.
-type CreateResult struct {
-	commonResult
-}
-
-// DeleteResult contains the response body and error from a Delete request.
-type DeleteResult struct {
-	golangsdk.ErrResult
-}
-
-// UpdateResult contains the response body and error from an Update request.
-type UpdateResult struct {
-	commonResult
 }
